requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
sqlalchemy>=2.0
python-multipart>=0.0.9
aiofiles>=23.2
pycryptodome>=3.20
pypdf>=4.2
python-docx>=1.1
rank-bm25>=0.2.2
jieba>=0.42.1
